requirements.txt
python-docx
requests
faiss-cpu
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.0
langchain_nvidia_ai_endpoints==0.3.10
streamlit==1.40.1
